+++ server/database/mockdb.go
@@ -5,25 +5,7 @@ import (
 	"github.com/YoshiRussell/bookclubapp/server/models"
 )
 
-type mockENV struct {
-	DB map[string]models.Book
-}
-
-
-func MockDatabaseENVInit() (Bookstore, error) {
-	bookstore := mockENV{
-		DB : make(map[string]models.Book),
-	}
-	bookstore.DB["testID"] = models.Book {
-		Isbn: "testIsbn", 
-		Title: "testTitle", 
-		Author: "testAuthor", 
-		Price: 69.99,
-	}
-	return &bookstore, nil
-}
-
-func (this *mockENV) GetALLBooks() ([]models.Book, error) {
+func (this *MockDb) GetALLBooks() ([]models.Book, error) {
 	bks := make([]models.Book, 0)
 	for _, v := range this.DB {
 		fmt.Println(v.Author)
@@ -31,7 +13,7 @@ func (this *mockENV) GetALLBooks() ([]models.Book, error) {
 	return bks, nil
 }
 
-func (this *mockENV) Close() {
+func (this *MockDb) Close() {
 	fmt.Println("closing mock database")
 }
